Decoupling set/add
