+++ src/commands/add.js
@@ -1,6 +1,7 @@
 const chalk = require('chalk');
 
 const {	SlashCommandBuilder } = require('@discordjs/builders')
+const querySet = require('../queries/set.js')
 const config = require('../config/config.json')
 const { add } = config
 
@@ -20,31 +21,17 @@ module.exports = {
 		const { reply } = add 
 		
 		if (quote.includes("<@!")) {
-			console.log(`Message contains mention, skipping`)
+			console.log(chalk.yellow(`Message contains mention, skipping`))
 			interaction.reply({content: reply.error.mention,ephemeral: true})
 		} else {
 			try {
-
-				const db = require('firebase-admin/firestore').getFirestore()
-				const collection = db.collection(process.env.COLLECTION_NAME)
-
-				let quotes = new Array()
-				const snapshot = await collection.get()
-				snapshot.forEach(doc => { quotes.push(doc.data().text) });
-				
-				if (quotes.includes(quote)) {
-					console.warn(chalk.yellow('Skipping quote insertion, already found in database'))
-					interaction.reply({content: reply.error.duplicate, ephemeral: true})
-				} else  {
-					const docRef = await collection.doc();
-					const now = new Date()
-					const data = { 'id': docRef.id, 'text': quote, 'time_added': now.getTime(), 'time_added_hr': now.toTimeString() }
-					await docRef.set(data);
-					console.log(`Added quote: ${quote}`)
-					console.log(`Associated document data: ${JSON.stringify(data)}`)
+				querySet.execute(quote).then(function (result) {
+					console.log(`Quote added successfully: ${result}`)
 					interaction.reply(`${reply.success}\n${quote}`)
-				}
-			
+				}).catch(function (error) {
+					console.log(chalk.red(error))
+					interaction.reply({content: reply.error.duplicate, ephemeral: true})
+				})	
 			} catch (error) {
 				console.log(chalk.red(error))
 				interaction.reply({ content: config.error_generic, ephemeral: true })
